# api/docs/api/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
    PlayerRole:
      enum:
        - Initiator
        - Duelist
        - Controller
        - Sentinel
        - Flex
        - IGL
      type: string
...
    TeamApiModel:
      properties:
        short_name:
          type: string
        full_name:
          type: string
        description:
          type: string
        country:
          type: string
        logo_image_file:
          type: string
        id:
          type: number
          format: double
      required:
        - short_name
        - full_name
        - description
        - country
      type: object
      additionalProperties: false
...
    ItemsWithPagination_TeamApiModel_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/TeamApiModel"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    TeamStats:
      properties:
        team:
          $ref: "#/components/schemas/TeamApiModel"
        tournamentsWon:
          type: number
          format: double
        tournamentsParticipated:
          type: number
          format: double
        winrate:
          type: number
          format: double
        totalMatchesPlayed:
          type: number
          format: double
        totalMatchesWon:
          type: number
          format: double
        totalMatchesLost:
          type: number
          format: double
        mapWinrate:
          type: number
          format: double
        totalMapsPlayed:
          type: number
          format: double
        totalMapsWon:
          type: number
          format: double
        totalMapsLost:
          type: number
          format: double
      required:
        - team
        - tournamentsWon
        - tournamentsParticipated
        - winrate
        - totalMatchesPlayed
        - totalMatchesWon
        - totalMatchesLost
        - mapWinrate
        - totalMapsPlayed
        - totalMapsWon
        - totalMapsLost
      type: object
      additionalProperties: false
    ItemsWithPagination_TeamStats_:
      properties:
        items:
          items:
            $ref: "#/components/schemas/TeamStats"
          type: array
        total:
          type: number
          format: double
      required:
        - items
        - total
      type: object
      additionalProperties: false
    PlayerAttributesApiModel:
      properties:
        clutch:
          type: number
          format: double
        awareness:
          type: number
          format: double
        aim:
          type: number
          format: double
        positioning:
          type: number
          format: double
        game_reading:
          type: number
          format: double
        resilience:
          type: number
          format: double
        confidence:
          type: number
          format: double
        strategy:
          type: number
          format: double
        adaptability:
          type: number
          format: double
        communication:
          type: number
          format: double
        unpredictability:
          type: number
          format: double
        game_sense:
          type: number
          format: double
        decision_making:
          type: number
          format: double
        rage_fuel:
          type: number
          format: double
        teamwork:
          type: number
          format: double
        utility_usage:
          type: number
          format: double
      required:
        - clutch
        - awareness
        - aim
        - positioning
        - game_reading
        - resilience
        - confidence
        - strategy
        - adaptability
        - communication
        - unpredictability
        - game_sense
        - decision_making
        - rage_fuel
        - teamwork
        - utility_usage
      type: object
      additionalProperties: false
    PlayerApiModel:
      properties:
        nickname:
          type: string
        full_name:
          type: string
        age:
          type: number
          format: double
        country:
          type: string
        team_id:
          type: number
          format: double
        role:
          $ref: "#/components/schemas/PlayerRole"
        player_attributes:
          $ref: "#/components/schemas/PlayerAttributesApiModel"
        id:
          type: number
          format: double
      required:
        - nickname
        - full_name
        - age
        - country
        - team_id
        - role
        - player_attributes
      type: object
      additionalProperties: false
    PlayerDuelResults:
      properties:
        winner:
          allOf:
            - $ref: "#/components/schemas/PlayerApiModel"
          nullable: true
        loser:
          allOf:
            - $ref: "#/components/schemas/PlayerApiModel"
          nullable: true
        startedTradeDuel:
          type: boolean
      required:
        - winner
        - loser
        - startedTradeDuel
      type: object
      additionalProperties: false
    RoundStateApiModel:
      properties:
        round:
          type: number
          format: double
        duel:
          $ref: "#/components/schemas/PlayerDuelResults"
        previous_duel:
          $ref: "#/components/schemas/PlayerDuelResults"
        team1_alive_players:
          items:
            $ref: "#/components/schemas/PlayerApiModel"
          type: array
        team2_alive_players:
          items:
            $ref: "#/components/schemas/PlayerApiModel"
          type: array
        team_won:
          allOf:
            - $ref: "#/components/schemas/TeamApiModel"
          nullable: true
        finished:
          type: boolean
      required:
        - round
        - duel
        - previous_duel
        - team1_alive_players
        - team2_alive_players
        - team_won
        - finished
      type: object
      additionalProperties: false
    Weapon:
      enum:
        - Vandal
        - Phantom
        - Operator
        - Sheriff
        - Ghost
        - Marshal
        - Ares
        - Odin
        - Bucky
        - Judge
        - Frenzy
        - Shorty
      type: string
    GameLogApiModel:
      properties:
        round_state:
          $ref: "#/components/schemas/RoundStateApiModel"
        duel_buff:
          type: number
          format: double
        trade_buff:
          type: number
          format: double
        trade:
          type: boolean
        weapon:
          $ref: "#/components/schemas/Weapon"
        game_id:
          type: number
          format: double
        team1_player_id:
          type: number
          format: double
        team2_player_id:
          type: number
          format: double
        player_killed_id:
          type: number
          format: double
        included_on_player_stats:
          type: boolean
        included_on_team_stats:
          type: boolean
        id:
          type: number
          format: double
      required:
        - round_state
        - duel_buff
        - trade_buff
        - trade
        - weapon
        - game_id
        - team1_player_id
        - team2_player_id
        - player_killed_id
        - included_on_player_stats
        - included_on_team_stats
      type: object
      additionalProperties: false
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
info:
  title: backend
  version: 0.1.0
  contact: {}
paths:
...
  /teams:
    get:
      operationId: getTeams
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_TeamApiModel_"
      description: Get all teams with optional filtering
      security: []
      parameters:
        - description: Optional country filter
          in: query
          name: country
          required: false
          schema:
            type: string
        - description: Maximum number of teams to return
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of teams to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
    post:
      operationId: createTeam
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamApiModel"
      description: Creates a new team
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                short_name:
                  type: string
                full_name:
                  type: string
                description:
                  type: string
                country:
                  type: string
                logo_image_file:
                  type: string
                  format: binary
              required:
                - short_name
                - full_name
                - description
                - country
  /teams/stats:
    get:
      operationId: getTeamsStats
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsWithPagination_TeamStats_"
      description: Retrieves stats for all teams
      security: []
      parameters:
        - description: Maximum number of teams to include
          in: query
          name: limit
          required: false
          schema:
            default: 10
            format: double
            type: number
        - description: Number of teams to skip
          in: query
          name: offset
          required: false
          schema:
            default: 0
            format: double
            type: number
  /teams/bulk:
    post:
      operationId: createTeamsBulk
      responses:
        "201":
          description: Teams created successfully
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/TeamApiModel"
                type: array
      description: Creates multiple teams from a bulk upload
      security: []
      parameters: []
      requestBody:
        description: Array of team data to create
        required: true
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/TeamApiModel"
              type: array
              description: Array of team data to create
  /teams/{teamId}:
    get:
      operationId: getTeam
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamApiModel"
      description: Get a specific team by ID
      security: []
      parameters:
        - description: The ID of the team to retrieve
          in: path
          name: teamId
          required: true
          schema:
            format: double
            type: number
    put:
      operationId: updateTeam
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamApiModel"
      description: Updates an existing team
      security: []
      parameters:
        - description: The ID of the team to update
          in: path
          name: teamId
          required: true
          schema:
            format: double
            type: number
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                short_name:
                  type: string
                full_name:
                  type: string
                description:
                  type: string
                country:
                  type: string
                logo_image_file:
                  type: string
                  format: binary
              required:
                - short_name
                - full_name
                - description
                - country
    delete:
      operationId: deleteTeam
      responses:
        "204":
          description: No content
      description: Deletes a team
      security: []
      parameters:
        - description: The ID of the team to delete
          in: path
          name: teamId
          required: true
          schema:
            format: double
            type: number
  /teams/{teamId}/stats:
    get:
      operationId: getTeamStats
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamStats"
      description: Retrieves stats for a specific team
      security: []
      parameters:
        - description: The ID of the team to retrieve stats for
          in: path
          name: teamId
          required: true
          schema:
            format: double
            type: number
  /teams/{teamId}/players:
    get:
      operationId: getTeamPlayers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PlayerApiModel"
                type: array
      description: Retrieves all players for a specific team
      security: []
      parameters:
        - description: The ID of the team to retrieve players for
          in: path
          name: teamId
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/rounds/{round}/play:
    post:
      operationId: playRound
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundStateApiModel"
      description: Play a full round in a game
      security: []
      parameters:
        - description: The ID of the game
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
        - description: The round number to play
          in: path
          name: round
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/rounds/{round}/duel:
    post:
      operationId: playDuel
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundStateApiModel"
      description: Play a single duel in a round
      security: []
      parameters:
        - description: The ID of the game
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
        - description: The round number
          in: path
          name: round
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/rounds/last/duel:
    get:
      operationId: getLastDuel
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameLogApiModel"
      description: Get the last duel in a game
      security: []
      parameters:
        - description: The ID of the game
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/rounds/last:
    get:
      operationId: getLastRound
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GameLogApiModel"
                type: array
      description: Get logs for the last round in a game
      security: []
      parameters:
        - description: The ID of the game
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
  /games/{gameId}/rounds/{round}:
    get:
      operationId: getRound
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/GameLogApiModel"
                type: array
      description: Get logs for a specific round in a game
      security: []
      parameters:
        - description: The ID of the game
          in: path
          name: gameId
          required: true
          schema:
            format: double
            type: number
        - description: The round number
          in: path
          name: round
          required: true
          schema:
            format: double
            type: number
...
servers:
  - url: /api